# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Warrant control plane
  version: 0.1.0
  summary: Issue scope-bound mandates, decide actions against them, and produce evidence anyone can verify offline.
  description: >
    Every response body that carries a signature is a Warrant format with its own version string,
    independent of this API's version. `/v1` may change while `warrant/evidence-pack/v0.1` does not,
    and the reverse is also true. A relying party should pin the format, not the endpoint.

    Authentication is declared, not guessed. A deployment runs in one of two modes and reports
    which on `/health` as `auth`. In `open` mode there is no authentication and anyone can issue
    mandates under the published demonstration keys, which exist so that a reader can reproduce and
    forge packs for themselves. In `required` mode the authority endpoints need a bearer token
    issued by the identity provider named in `authIssuer`.

    Authentication decides who may call, never what the answer is. The verdict on an action is a
    function of the mandate chain and the signed evaluation inputs alone, so an offline verifier
    that has never seen the caller reproduces the same result. No endpoint below changes behaviour
    based on who is logged in.

    The verification endpoints and the demonstration scenarios are public in both modes. A relying
    party checking evidence must never need an account here.
servers:
  - url: /v1

tags:
  - name: mandates
    description: Issuance, delegation and revocation of authority.
  - name: actions
    description: Deciding an action against the authority presented with it.
  - name: organisations
    description: Who may act under which organisation, and in what role.
  - name: verification
    description: Relying-party operations. Nothing here requires trusting this service.

paths:
  /mandates:
    post:
      tags: [mandates]
      summary: Issue a root mandate
      description: The issuer must be a principal key. Authority granted here is the ceiling for every hop below it.
      security: [{ bearerAuth: [] }]
      responses:
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "201": { description: The signed mandate, at depth 0. }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "422": { description: The requested scope is not one this issuer may grant. }

  /mandates/{id}:
    get:
      tags: [mandates]
      summary: Read a mandate and the chain above it
      parameters: [{ $ref: "#/components/parameters/MandateId" }]
      security: [{ bearerAuth: [] }]
      responses:
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "200": { description: The mandate and its ancestors, root first. }
        "404": { $ref: "#/components/responses/NotFound" }

  /mandates/{id}/delegations:
    post:
      tags: [mandates]
      summary: Delegate a narrower mandate
      description: >
        Refused if the delegation would widen any dimension of scope, exceed the parent's validity
        window, or pass the maximum delegation depth. Absent fields inherit from the parent; they
        never widen.
      parameters: [{ $ref: "#/components/parameters/MandateId" }]
      security: [{ bearerAuth: [] }]
      responses:
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "201": { description: The signed child mandate. }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: >
            `delegation_would_widen`, naming the dimension that failed, for example
            `scope/per_action_limit_exceeded`.

  /mandates/{id}/revocation:
    post:
      tags: [mandates]
      summary: Revoke a mandate
      description: Revocation is published as a signed snapshot carrying the moment it was taken.
      parameters: [{ $ref: "#/components/parameters/MandateId" }]
      security: [{ bearerAuth: [] }]
      responses:
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "201": { description: The updated signed revocation snapshot. }
        "404": { $ref: "#/components/responses/NotFound" }

  /actions:
    post:
      tags: [actions]
      summary: Present an action for decision
      description: >
        Runs the full check pipeline and returns a signed decision plus an evidence pack. **A refused
        action is a 201, not an error** — a BLOCK is a successful decision and valid evidence. Only a
        malformed request is a 4xx.
      security: [{ bearerAuth: [] }]
      responses:
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "201":
          description: The signed decision and its evidence pack. Verdict is ALLOW, BLOCK or ESCALATE.
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "404": { $ref: "#/components/responses/NotFound" }

  /checkpoint:
    post:
      tags: [verification]
      summary: Take a signed checkpoint over the current ledger head
      description: >
        On demand only. Nothing schedules this. The checkpoint commits to the ledger's head digest
        and entry count, and is what would be submitted to an external transparency log — but
        anchoring is supplemental. **A checkpoint that was never anchored is still a valid signed
        commitment, and no authorization decision or verification depends on a log being reachable.**
      security: [{ bearerAuth: [] }]
      responses:
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "201": { description: The signed checkpoint. }
        "422": { description: "`ledger_empty` — there is nothing to commit to yet." }

  /organisations:
    post:
      tags: [organisations]
      summary: Create an organisation and become its owner
      description: >
        Self-service. The authenticated caller becomes the owner. Authority is always recorded under
        an organisation, so an account that belongs to none cannot issue a mandate.
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: The organisation, with the caller's role. }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422": { description: "`organisation_exists`." }
    get:
      tags: [organisations]
      summary: List the organisations this account belongs to
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: The caller's organisations and role in each. }
        "401": { $ref: "#/components/responses/Unauthenticated" }

  /organisations/{id}/members:
    get:
      tags: [organisations]
      summary: List the members of an organisation
      description: Scoped to the caller's own organisation; another organisation reads as absent.
      parameters: [{ $ref: "#/components/parameters/OrganisationId" }]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: The members and their roles. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [organisations]
      summary: Grant a role to an identity
      description: >
        Needs `admin` or `owner`. The identity is named by its subject at the configured provider,
        because this service cannot enumerate that provider's users.
      parameters: [{ $ref: "#/components/parameters/OrganisationId" }]
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: The membership that now exists. }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/InsufficientRole" }
        "404": { $ref: "#/components/responses/NotFound" }

  /organisations/{id}/members/{accountId}/withdrawal:
    post:
      tags: [organisations]
      summary: Withdraw a membership
      description: Needs `admin` or `owner`. An account may not withdraw its own membership.
      parameters:
        - { $ref: "#/components/parameters/OrganisationId" }
        - { $ref: "#/components/parameters/AccountId" }
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: The membership is gone. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/InsufficientRole" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { description: "`last_owner`." }

  /trust-roots:
    get:
      tags: [verification]
      summary: The public keys this deployment signs with
      description: >
        Public halves only. Keys carry an optional signing lifecycle; a key outside its signing
        window cannot sign new artifacts but everything it already signed stays verifiable.
      responses:
        "200": { description: An array of trust roots. Never contains private key material. }

  /verify:
    post:
      tags: [verification]
      summary: Verify an evidence pack
      description: >
        Offered as a convenience. It is not the trustworthy way to verify — a relying party should
        run the standalone verifier against keys it obtained itself, because asking the issuer whether
        its own evidence is good proves nothing.
      responses:
        "200":
          description: >
            A verification report. `result` is VERIFIED or INVALID and answers whether the evidence
            is genuine; `authority` explains the verdict and is a finding, not a fault.
        "400": { $ref: "#/components/responses/InvalidRequest" }

  /evidence/{id}:
    get:
      tags: [verification]
      summary: Fetch a stored evidence pack
      parameters: [{ $ref: "#/components/parameters/PackId" }]
      responses:
        "200": { description: The evidence pack. }
        "404": { $ref: "#/components/responses/NotFound" }

  /scenarios:
    get:
      tags: [verification]
      summary: List the deterministic demonstration scenarios
      responses:
        "200": { description: Scenario identifiers, questions and expected verdicts. }

  /scenarios/{id}:
    get:
      tags: [verification]
      summary: Read one demonstration scenario
      parameters: [{ $ref: "#/components/parameters/ScenarioId" }]
      responses:
        "200": { description: The scenario with its chain, request, decision and pack. }
        "404": { $ref: "#/components/responses/NotFound" }

components:
  parameters:
    MandateId:
      name: id
      in: path
      required: true
      schema: { type: string }
    PackId:
      name: id
      in: path
      required: true
      schema: { type: string }
    ScenarioId:
      name: id
      in: path
      required: true
      schema: { type: string }
    OrganisationId:
      name: id
      in: path
      required: true
      schema: { type: string }
    AccountId:
      name: accountId
      in: path
      required: true
      schema: { type: string }

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        An ES256 access token from the configured identity provider, checked against its published
        JWKS. Only ES256 is accepted, so a public key cannot be replayed as an HMAC secret. Claims
        under `user_metadata` are user-writable and are never read.

  responses:
    InvalidRequest:
      description: "`invalid_request`, with a per-field list of what failed."
    NotFound:
      description: "`not_found`."
    Unauthenticated:
      description: "`unauthenticated`, with a `WWW-Authenticate: Bearer` header. Only in `required` mode."
    InsufficientRole:
      description: "`insufficient_role`, naming the role the action needs and the one the account holds."